feat: 🎸 Use artifact server bundled carvel toolset
To make sure we are able to produce reproducible builds, we need to make sure the carvel toolset installed remains same for a particular version. Currently carvel toolset is installed using upstream install script from main branch which will install the latest at the moment of time which can be different depending upon the time of building the image. To avoid this, now the artifact server includes the carvel toolset bundled.

Signed-off-by: Sriraman S <sriraman.iyengar@broadcom.com>