Enable readonly access to localhost docker distribution registry
This fixes the unauthenticated access to localhost docker distribution registry by enabling the readonly flag in the registry config. Additionally, this removes the delete flag from the config to seal the registry. Additionally, it updates the legacy static pod config to take the updated sealed registry config. In the retag images operation, it ensures that the push to registry option is disabled.

Signed-off-by: Sriraman S <sriraman.iyengar@broadcom.com>