fix: 🐛 Update the cloud-init to go older netplan route
Ubuntu 24 with netplan and Cloud-Init 24.2+ is not compatible with vmoperator shipped with VCF9.0 .This is because this cloud-init version uses netplan module if available to pass the network config that is passed by vm operator. and this network config when having a mtu:0 value gets parsed into mtu:null which causes the cloud-init script to fail.

Signed-off-by: Sriraman S <sriraman.iyengar@broadcom.com>